# Add scheduler plugin execution duration metric.