--- a/pkg/scheduler/framework/v1alpha1/cycle_state.go
+++ b/pkg/scheduler/framework/v1alpha1/cycle_state.go
@@ -44,6 +44,8 @@ type StateKey string
 type CycleState struct {
 	mx      sync.RWMutex
 	storage map[StateKey]StateData
+	// if recordFrameworkMetrics is true, framework metrics will be recorded for this cycle.
+	recordFrameworkMetrics bool
 }
 
 // NewCycleState initializes a new CycleState and returns its pointer.
@@ -51,6 +53,22 @@ func NewCycleState() *CycleState {
 	return &CycleState{
 		storage: make(map[StateKey]StateData),
 	}
+}
+
+// ShouldRecordFrameworkMetrics returns whether framework metrics should be recorded.
+func (c *CycleState) ShouldRecordFrameworkMetrics() bool {
+	if c == nil {
+		return false
+	}
+	return c.recordFrameworkMetrics
+}
+
+// SetRecordFrameworkMetrics sets recordFrameworkMetrics to the given value.
+func (c *CycleState) SetRecordFrameworkMetrics(flag bool) {
+	if c == nil {
+		return
+	}
+	c.recordFrameworkMetrics = flag
 }
 
 // Clone creates a copy of CycleState and returns its pointer. Clone returns