--- a/pkg/scheduler/framework/v1alpha1/framework_test.go
+++ b/pkg/scheduler/framework/v1alpha1/framework_test.go
@@ -133,6 +133,17 @@ type TestPlugin struct {
 	inj  injectedResult
 }
 
+type TestPluginPreFilterExtension struct {
+	inj injectedResult
+}
+
+func (e *TestPluginPreFilterExtension) AddPod(ctx context.Context, state *CycleState, podToSchedule *v1.Pod, podToAdd *v1.Pod, nodeInfo *schedulernodeinfo.NodeInfo) *Status {
+	return NewStatus(Code(e.inj.PreFilterAddPodStatus), "injected status")
+}
+func (e *TestPluginPreFilterExtension) RemovePod(ctx context.Context, state *CycleState, podToSchedule *v1.Pod, podToRemove *v1.Pod, nodeInfo *schedulernodeinfo.NodeInfo) *Status {
+	return NewStatus(Code(e.inj.PreFilterRemovePodStatus), "injected status")
+}
+
 func (pl *TestPlugin) Name() string {
 	return pl.name
 }
@@ -150,7 +161,7 @@ func (pl *TestPlugin) PreFilter(ctx context.Context, state *CycleState, p *v1.Po
 }
 
 func (pl *TestPlugin) PreFilterExtensions() PreFilterExtensions {
-	return nil
+	return &TestPluginPreFilterExtension{inj: pl.inj}
 }
 
 func (pl *TestPlugin) Filter(ctx context.Context, state *CycleState, pod *v1.Pod, nodeInfo *schedulernodeinfo.NodeInfo) *Status {
@@ -586,6 +597,7 @@ func TestPreFilterPlugins(t *testing.T) {
 }
 
 func TestRecordingMetrics(t *testing.T) {
+	state := &CycleState{recordFrameworkMetrics: true}
 	tests := []struct {
 		name               string
 		action             func(f Framework)
@@ -595,117 +607,149 @@ func TestRecordingMetrics(t *testing.T) {
 	}{
 		{
 			name:               "PreFilter - Success",
-			action:             func(f Framework) { f.RunPreFilterPlugins(context.Background(), nil, pod) },
+			action:             func(f Framework) { f.RunPreFilterPlugins(context.Background(), state, pod) },
+			wantExtensionPoint: "PreFilter",
+			wantStatus:         Success,
+		},
+		{
+			name:               "PreFilterAddPod - Success",
+			action:             func(f Framework) { f.RunPreFilterExtensionAddPod(context.Background(), state, pod, nil, nil) },
+			wantExtensionPoint: "PreFilterExtensionAddPod",
+			wantStatus:         Success,
+		},
+		{
+			name:               "PreFilterRemovePod - Success",
+			action:             func(f Framework) { f.RunPreFilterExtensionRemovePod(context.Background(), state, pod, nil, nil) },
+			wantExtensionPoint: "PreFilterExtensionRemovePod",
+			wantStatus:         Success,
+		},
+		{
+			name:               "PreFilterRemovePod - Success",
+			action:             func(f Framework) { f.RunPreFilterPlugins(context.Background(), state, pod) },
 			wantExtensionPoint: "PreFilter",
 			wantStatus:         Success,
 		},
 		{
 			name:               "Filter - Success",
-			action:             func(f Framework) { f.RunFilterPlugins(context.Background(), nil, pod, nil) },
+			action:             func(f Framework) { f.RunFilterPlugins(context.Background(), state, pod, nil) },
 			wantExtensionPoint: "Filter",
 			wantStatus:         Success,
 		},
 		{
 			name:               "PostFilter - Success",
-			action:             func(f Framework) { f.RunPostFilterPlugins(context.Background(), nil, pod, nil, nil) },
+			action:             func(f Framework) { f.RunPostFilterPlugins(context.Background(), state, pod, nil, nil) },
 			wantExtensionPoint: "PostFilter",
 			wantStatus:         Success,
 		},
 		{
 			name:               "Score - Success",
-			action:             func(f Framework) { f.RunScorePlugins(context.Background(), nil, pod, nodes) },
+			action:             func(f Framework) { f.RunScorePlugins(context.Background(), state, pod, nodes) },
 			wantExtensionPoint: "Score",
 			wantStatus:         Success,
 		},
 		{
 			name:               "Reserve - Success",
-			action:             func(f Framework) { f.RunReservePlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunReservePlugins(context.Background(), state, pod, "") },
 			wantExtensionPoint: "Reserve",
 			wantStatus:         Success,
 		},
 		{
 			name:               "Unreserve - Success",
-			action:             func(f Framework) { f.RunUnreservePlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunUnreservePlugins(context.Background(), state, pod, "") },
 			wantExtensionPoint: "Unreserve",
 			wantStatus:         Success,
 		},
 		{
 			name:               "PreBind - Success",
-			action:             func(f Framework) { f.RunPreBindPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunPreBindPlugins(context.Background(), state, pod, "") },
 			wantExtensionPoint: "PreBind",
 			wantStatus:         Success,
 		},
 		{
 			name:               "Bind - Success",
-			action:             func(f Framework) { f.RunBindPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunBindPlugins(context.Background(), state, pod, "") },
 			wantExtensionPoint: "Bind",
 			wantStatus:         Success,
 		},
 		{
 			name:               "PostBind - Success",
-			action:             func(f Framework) { f.RunPostBindPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunPostBindPlugins(context.Background(), state, pod, "") },
 			wantExtensionPoint: "PostBind",
 			wantStatus:         Success,
 		},
 		{
 			name:               "Permit - Success",
-			action:             func(f Framework) { f.RunPermitPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunPermitPlugins(context.Background(), state, pod, "") },
 			wantExtensionPoint: "Permit",
 			wantStatus:         Success,
 		},
 
 		{
 			name:               "PreFilter - Error",
-			action:             func(f Framework) { f.RunPreFilterPlugins(context.Background(), nil, pod) },
+			action:             func(f Framework) { f.RunPreFilterPlugins(context.Background(), state, pod) },
 			inject:             injectedResult{PreFilterStatus: int(Error)},
 			wantExtensionPoint: "PreFilter",
 			wantStatus:         Error,
 		},
 		{
+			name:               "PreFilterAddPod - Error",
+			action:             func(f Framework) { f.RunPreFilterExtensionAddPod(context.Background(), state, pod, nil, nil) },
+			inject:             injectedResult{PreFilterAddPodStatus: int(Error)},
+			wantExtensionPoint: "PreFilterExtensionAddPod",
+			wantStatus:         Error,
+		},
+		{
+			name:               "PreFilterRemovePod - Error",
+			action:             func(f Framework) { f.RunPreFilterExtensionRemovePod(context.Background(), state, pod, nil, nil) },
+			inject:             injectedResult{PreFilterRemovePodStatus: int(Error)},
+			wantExtensionPoint: "PreFilterExtensionRemovePod",
+			wantStatus:         Error,
+		},
+		{
 			name:               "Filter - Error",
-			action:             func(f Framework) { f.RunFilterPlugins(context.Background(), nil, pod, nil) },
+			action:             func(f Framework) { f.RunFilterPlugins(context.Background(), state, pod, nil) },
 			inject:             injectedResult{FilterStatus: int(Error)},
 			wantExtensionPoint: "Filter",
 			wantStatus:         Error,
 		},
 		{
 			name:               "PostFilter - Error",
-			action:             func(f Framework) { f.RunPostFilterPlugins(context.Background(), nil, pod, nil, nil) },
+			action:             func(f Framework) { f.RunPostFilterPlugins(context.Background(), state, pod, nil, nil) },
 			inject:             injectedResult{PostFilterStatus: int(Error)},
 			wantExtensionPoint: "PostFilter",
 			wantStatus:         Error,
 		},
 		{
 			name:               "Score - Error",
-			action:             func(f Framework) { f.RunScorePlugins(context.Background(), nil, pod, nodes) },
+			action:             func(f Framework) { f.RunScorePlugins(context.Background(), state, pod, nodes) },
 			inject:             injectedResult{ScoreStatus: int(Error)},
 			wantExtensionPoint: "Score",
 			wantStatus:         Error,
 		},
 		{
 			name:               "Reserve - Error",
-			action:             func(f Framework) { f.RunReservePlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunReservePlugins(context.Background(), state, pod, "") },
 			inject:             injectedResult{ReserveStatus: int(Error)},
 			wantExtensionPoint: "Reserve",
 			wantStatus:         Error,
 		},
 		{
 			name:               "PreBind - Error",
-			action:             func(f Framework) { f.RunPreBindPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunPreBindPlugins(context.Background(), state, pod, "") },
 			inject:             injectedResult{PreBindStatus: int(Error)},
 			wantExtensionPoint: "PreBind",
 			wantStatus:         Error,
 		},
 		{
 			name:               "Bind - Error",
-			action:             func(f Framework) { f.RunBindPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunBindPlugins(context.Background(), state, pod, "") },
 			inject:             injectedResult{BindStatus: int(Error)},
 			wantExtensionPoint: "Bind",
 			wantStatus:         Error,
 		},
 		{
 			name:               "Permit - Error",
-			action:             func(f Framework) { f.RunPermitPlugins(context.Background(), nil, pod, "") },
+			action:             func(f Framework) { f.RunPermitPlugins(context.Background(), state, pod, "") },
 			inject:             injectedResult{PermitStatus: int(Error)},
 			wantExtensionPoint: "Permit",
 			wantStatus:         Error,
@@ -714,6 +758,10 @@ func TestRecordingMetrics(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			metrics.Register()
+			metrics.FrameworkExtensionPointDuration.Reset()
+			metrics.PluginExecutionDuration.Reset()
+
 			plugin := &TestPlugin{name: testPlugin, inj: tt.inject}
 			r := make(Registry)
 			r.Register(testPlugin,
@@ -733,16 +781,22 @@ func TestRecordingMetrics(t *testing.T) {
 				PostBind:   pluginSet,
 				Unreserve:  pluginSet,
 			}
-			f, err := NewFramework(r, plugins, emptyArgs)
+			recorder := newMetricsRecorder(100, time.Nanosecond)
+			f, err := NewFramework(r, plugins, emptyArgs, withMetricsRecorder(recorder))
 			if err != nil {
 				t.Fatalf("Failed to create framework for testing: %v", err)
 			}
-			metrics.Register()
-			metrics.FrameworkExtensionPointDuration.Reset()
 
 			tt.action(f)
 
+			// Stop the goroutine which records metrics and ensure it's stopped.
+			close(recorder.stopCh)
+			<-recorder.isStoppedCh
+			// Try to clean up the metrics buffer again in case it's not empty.
+			recorder.flushMetrics()
+
 			collectAndCompareFrameworkMetrics(t, tt.wantExtensionPoint, tt.wantStatus)
+			collectAndComparePluginMetrics(t, tt.wantExtensionPoint, testPlugin, tt.wantStatus)
 		})
 	}
 }
@@ -765,6 +819,9 @@ func TestPermitWaitingMetric(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
+			metrics.Register()
+			metrics.PermitWaitDuration.Reset()
+
 			plugin := &TestPlugin{name: testPlugin, inj: tt.inject}
 			r := make(Registry)
 			r.Register(testPlugin,
@@ -778,8 +835,6 @@ func TestPermitWaitingMetric(t *testing.T) {
 			if err != nil {
 				t.Fatalf("Failed to create framework for testing: %v", err)
 			}
-			metrics.Register()
-			metrics.PermitWaitDuration.Reset()
 
 			f.RunPermitPlugins(context.TODO(), nil, pod, "")
 
@@ -839,17 +894,19 @@ func buildScoreConfigWithWeights(weights map[string]int32, ps ...string) *config
 }
 
 type injectedResult struct {
-	ScoreRes         int64 `json:"scoreRes,omitempty"`
-	NormalizeRes     int64 `json:"normalizeRes,omitempty"`
-	ScoreStatus      int   `json:"scoreStatus,omitempty"`
-	NormalizeStatus  int   `json:"normalizeStatus,omitempty"`
-	PreFilterStatus  int   `json:"preFilterStatus,omitempty"`
-	FilterStatus     int   `json:"filterStatus,omitempty"`
-	PostFilterStatus int   `json:"postFilterStatus,omitempty"`
-	ReserveStatus    int   `json:"reserveStatus,omitempty"`
-	PreBindStatus    int   `json:"preBindStatus,omitempty"`
-	BindStatus       int   `json:"bindStatus,omitempty"`
-	PermitStatus     int   `json:"permitStatus,omitempty"`
+	ScoreRes                 int64 `json:"scoreRes,omitempty"`
+	NormalizeRes             int64 `json:"normalizeRes,omitempty"`
+	ScoreStatus              int   `json:"scoreStatus,omitempty"`
+	NormalizeStatus          int   `json:"normalizeStatus,omitempty"`
+	PreFilterStatus          int   `json:"preFilterStatus,omitempty"`
+	PreFilterAddPodStatus    int   `json:"preFilterAddPodStatus,omitempty"`
+	PreFilterRemovePodStatus int   `json:"preFilterRemovePodStatus,omitempty"`
+	FilterStatus             int   `json:"filterStatus,omitempty"`
+	PostFilterStatus         int   `json:"postFilterStatus,omitempty"`
+	ReserveStatus            int   `json:"reserveStatus,omitempty"`
+	PreBindStatus            int   `json:"preBindStatus,omitempty"`
+	BindStatus               int   `json:"bindStatus,omitempty"`
+	PermitStatus             int   `json:"permitStatus,omitempty"`
 }
 
 func setScoreRes(inj injectedResult) (int64, *Status) {
@@ -869,6 +926,33 @@ func injectNormalizeRes(inj injectedResult, scores NodeScoreList) *Status {
 	return nil
 }
 
+func collectAndComparePluginMetrics(t *testing.T, wantExtensionPoint, wantPlugin string, wantStatus Code) {
+	m := collectHistogramMetric(metrics.PluginExecutionDuration)
+	if len(m.Label) != 3 {
+		t.Fatalf("Unexpected number of label pairs, got: %v, want: 2", len(m.Label))
+	}
+
+	if *m.Label[0].Value != wantExtensionPoint {
+		t.Errorf("Unexpected extension point label, got: %q, want %q", *m.Label[0].Value, wantExtensionPoint)
+	}
+
+	if *m.Label[1].Value != wantPlugin {
+		t.Errorf("Unexpected plugin label, got: %q, want %q", *m.Label[1].Value, wantPlugin)
+	}
+
+	if *m.Label[2].Value != wantStatus.String() {
+		t.Errorf("Unexpected status code label, got: %q, want %q", *m.Label[2].Value, wantStatus)
+	}
+
+	if *m.Histogram.SampleCount == 0 {
+		t.Error("Expect at least 1 sample")
+	}
+
+	if *m.Histogram.SampleSum <= 0 {
+		t.Errorf("Expect latency to be greater than 0, got: %v", *m.Histogram.SampleSum)
+	}
+}
+
 func collectAndCompareFrameworkMetrics(t *testing.T, wantExtensionPoint string, wantStatus Code) {
 	m := collectHistogramMetric(metrics.FrameworkExtensionPointDuration)
 
@@ -885,11 +969,11 @@ func collectAndCompareFrameworkMetrics(t *testing.T, wantExtensionPoint string, 
 	}
 
 	if *m.Histogram.SampleCount != 1 {
-		t.Errorf("Expect 1 sample, got: %v", m.Histogram.SampleCount)
+		t.Errorf("Expect 1 sample, got: %v", *m.Histogram.SampleCount)
 	}
 
 	if *m.Histogram.SampleSum <= 0 {
-		t.Errorf("Expect latency to be greater than 0, got: %v", m.Histogram.SampleSum)
+		t.Errorf("Expect latency to be greater than 0, got: %v", *m.Histogram.SampleSum)
 	}
 }
 
@@ -911,17 +995,17 @@ func collectAndComparePermitWaitDuration(t *testing.T, wantRes string) {
 		}
 
 		if *m.Histogram.SampleCount != 1 {
-			t.Errorf("Expect 1 sample, got: %v", m.Histogram.SampleCount)
+			t.Errorf("Expect 1 sample, got: %v", *m.Histogram.SampleCount)
 		}
 
 		if *m.Histogram.SampleSum <= 0 {
-			t.Errorf("Expect latency to be greater than 0, got: %v", m.Histogram.SampleSum)
+			t.Errorf("Expect latency to be greater than 0, got: %v", *m.Histogram.SampleSum)
 		}
 	}
 }
 
 func collectHistogramMetric(metric prometheus.Collector) *dto.Metric {
-	ch := make(chan prometheus.Metric, 1)
+	ch := make(chan prometheus.Metric, 100)
 	metric.Collect(ch)
 	select {
 	case got := <-ch: