<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**
> Uncomment only one ` /kind <>` line, hit enter to put that in a new line, and remove leading whitespace from that line:
>
> /kind api-change
> /kind bug
> /kind cleanup
> /kind design
> /kind documentation
> /kind failing-test
> /kind feature
> /kind flake

/kind feature

**What this PR does / why we need it**:
Helps us understand feature adoption as well as performance of each plugins.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #83416

**Special notes for your reviewer**:
I compared benchmark results for 
a) master branch, 
b) master branch with extension point metrics removed, 
c) this change and 
d) this change with sampling every 15 scheduling cycles.

a) vs b): b) improves over a) slightly in many cases, but almost negligible considering noises. 
c) vs d): By sampling the performance could be improved for some cases but not for all.
 
It's surprising that c) is better than a)/b) in test case #1,3,8 consistently, while worse in #2,4,5,6 and very close for #7.  I don't know how these numbers map to real use cases, given the value of getting visibility into each plugin, I prefer to get this change out and get some feedback from users.

Benchmark Result (ns/op)
--
Test # |  Test | Master | No metric | Periodic flush | Periodic flush with 1/15 sample
-- | -- | -- | -- | -- | --
1 | BenchmarkSchedulingV/5000Nodes/1000Pods-12 | 9237142 | 9229039 | 7761435 | 6754970
2 | BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12 | 9114743 | 8970976 | 10910383 | 9660827
3 | BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12 | 8656598 | 10362103 | 8220864 | 7031266
4 | BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12 | 14089110 | 13623460 | 15478250 | 14093959
5 | BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12 | 12615123 | 12240668 | 14704521 | 13194460
6 | BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12 | 11570522 | 11356818 | 13410923 | 11998329
7 | BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12 | 24290660 | 25244090 | 24926082 | 25813202
8 | BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12 | 19159768 | 18998646 | 17467853 | 18523843




**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
Add plugin_execution_duration_seconds metric for scheduler framework plugins.
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```