@liu-cong 2019-10-29
@ahg-g 
Added latency metric for Filter and Score plugins for doing perf tests. I couldn't run the perf tests locally for some reason. While I am trying to fix that, do you mind checking out this diff and run it on you machine?