@ahg-g 2019-11-11
Here are the benchmark numbers I am getting:

With the change
```
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           2237820 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           5520113 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           2506520 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           6115268 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           6025317 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           6197104 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000           7251515 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           3317381 ns/op
```

With the change:

```
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           2724880 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           5885653 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           2858365 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           6258707 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           6306560 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           6346709 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000           7600392 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           3556736 ns/op
```