@liu-cong 2019-11-11
I updated to the latest master and here are the new results:
```
make test-integration WHAT=./test/integration/scheduler_perf KUBE_TEST_VMODULE="''" KUBE_TEST_ARGS="-alsologtostderr=false -logtostderr=false -run=^$$ -bench=BenchmarkScheduling/5000Nodes/1000Pods -cpuprofile ~/cpu-master.out" |grep Benchmark
```

Master:
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           4900572 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           5321680 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           4943341 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           4078854 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           4590515 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           5134987 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000          12269336 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           6553250 ns/op

With the change:
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           4119963 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           6134407 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           4521283 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           4110121 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           4280762 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           3930512 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000          12414046 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           6739703 ns/op

I am still seeing better results with this change for the first test case, which is strange. Now I trust your results more than mine... Can you please take the latest commit with sampling and run it again?
@ahg-g