@ahg-g 2019-11-11
with the latest patch:

```
BenchmarkSchedulingV/5000Nodes/1000Pods-12          1000           2417580 ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12                    1000           5496006 ns/op
BenchmarkSchedulingSecrets/5000Nodes/1000Pods-12                            1000           2535986 ns/op
BenchmarkSchedulingInTreePVs/5000Nodes/1000Pods-12                          1000           6165100 ns/op
BenchmarkSchedulingMigratedInTreePVs/5000Nodes/1000Pods-12                  1000           6099270 ns/op
BenchmarkSchedulingCSIPVs/5000Nodes/1000Pods-12                             1000           6150523 ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/1000Pods-12                        1000           7309418 ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/1000Pods-12                       1000           3344050 ns/op

```